Aims to fix issue #11
Aims to fix issue #11, adds intelligent feature to remove all player targets is "Players" was already targeted in the array, and same for GM targets if "GM" or "DM" was already targeted.  Edge-casey.  But it felt weird targeting "Players" and then having this script offer up everyone's name still as a possible target.