allows up/down keydown handler to be bound just once, globally
this pulls the chat textarea's keydown.menufocus out of the nest function and allows it to be bound *just once* globally, instead of binding & unbinding repeatedly as whispers are detected / completed.  it allows our up/down key listener to step in front of the original FVTT keydown handler, but allows it to fall through to the original handler if our conditions are not met.